refactor: add ChevronDialog
