+++ lib/features/common/presentation/widgets/chevron_dialog.dart
@@ -1,0 +1,43 @@
+import 'package:auto_route/auto_route.dart';
+import 'package:flutter/material.dart';
+import 'package:repaint_mobile/features/common/presentation/widgets/flat_icon_button.dart';
+
+class ChevronDialog extends StatelessWidget {
+  const ChevronDialog({
+    super.key,
+    required this.children,
+  });
+
+  final List<Widget> children;
+
+  @override
+  Widget build(BuildContext context) {
+    return Dialog(
+      alignment: Alignment.center,
+      shape: const RoundedRectangleBorder(
+        borderRadius: BorderRadius.all(Radius.circular(8.0)),
+      ),
+      elevation: 0.0,
+      child: Padding(
+        padding: const EdgeInsets.symmetric(vertical: 24.0, horizontal: 16.0),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                FlatIconButton(
+                  onPressed: context.popRoute,
+                  icon: Icons.chevron_left,
+                ),
+                const Spacer()
+              ],
+            ),
+            const SizedBox(height: 24.0),
+            ...children,
+          ],
+        ),
+      ),
+    );
+  }
+}
